Inline unimportant extend
